make alphashape return WKT, instead of a set of vertices.
This is necessary for trying to return multipolygons without going crazy. Instead of trying to return vertices which will just have to be looped through in PL/PSQL,
we can just return a text object that we can pass through geometry().

In the future, we may be able to simplify this to just returning a straight geometry object. one thing at a time.

this greatly simplifies the alphashape() command